Add initial TUI app
Using Textual for the first time so there's definitely a lot that can be
improved. I'm sure that the complete app logic being crammed in a
150-line class is not the way to go.
A reasonable start would be to break up some of the logic into compound
widgets and also extract the math/science part away from the UI
composing code.

I'm also not a huge fan of the separate CSS file lying around...